fixed parsing error on device with dimensions not integer

+++ android/app/src/main/java/com/reactnativenavigation/params/parsers/OverlayParamsParser.java
@@ -24,8 +24,8 @@ public class OverlayParamsParser extends Parser {
 
         Bundle position = params.getBundle(POSITION);
 
-        result.top = position.getInt(POSITION_TOP);
-        result.left = position.getInt(POSITION_LEFT);
+        result.top = (int) position.getDouble(POSITION_TOP);
+        result.left = (int) position.getDouble(POSITION_LEFT);
         result.width = position.getInt(POSITION_WIDTH);
         result.height = position.getInt(POSITION_HEIGHT);
